FIX: image grids and its container.
Uses a different container for the content's
content (content-container). This allows images grids inside the
content-container to be scaled to content-container's size instead of
the regular container, which is not 100% width.

The whole reason for this fix is because using "width: 100vw;" for the
image grids created a horizontal scroll bar whereas "width: 100%;"
does not.